Tratamento de erro 404

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -18,6 +18,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos")
@@ -59,10 +60,12 @@ public class TopicosController {
 
     @GetMapping("/{id}")
 
-    public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-        Topico topico = topicoRepository.getOne(id);
-        return new DetalhesDoTopicoDto(topico);
-
+    public ResponseEntity <DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+        Optional <Topico> topico = topicoRepository.findById(id);
+        if (topico.isPresent()) {
+            return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+        }
+        return ResponseEntity.notFound().build();
 
     }
 
@@ -70,9 +73,13 @@ public class TopicosController {
     @Transactional
     public ResponseEntity <TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
 
-        Topico topico = form.atualizar(id, topicoRepository);
+        Optional <Topico> optional= topicoRepository.findById(id);
+        if (optional.isPresent()) {
+            Topico topico = form.atualizar(id, topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
+        return ResponseEntity.notFound().build();
 
-        return ResponseEntity.ok(new TopicoDto(topico));
     }
 
 
@@ -80,11 +87,17 @@ public class TopicosController {
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id) {
 
-            topicoRepository.deleteById(id);
+        Optional <Topico> optional= topicoRepository.findById(id);
+        if (optional.isPresent()) {
 
+            topicoRepository.deleteById(id);
             return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
+
 
     }
+
 
 }
